move custom recipes into BismuthRecipeManager
