+++ src/main/java/si/bismuth/BismuthServer.java
@@ -1,17 +1,8 @@
 package si.bismuth;
 
-import com.google.common.collect.Lists;
-import net.minecraft.crafting.CraftingManager;
-import net.minecraft.crafting.recipe.Ingredient;
-import net.minecraft.crafting.recipe.Recipe;
-import net.minecraft.crafting.recipe.ShapelessRecipe;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.nbt.NbtCompound;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.dedicated.DedicatedServer;
 import net.minecraft.server.entity.living.player.ServerPlayerEntity;
-import net.minecraft.util.DefaultedList;
 import net.minecraft.world.GameMode;
 import net.ornithemc.osl.entrypoints.api.ModInitializer;
 import net.ornithemc.osl.lifecycle.api.server.MinecraftServerEvents;
@@ -32,27 +23,6 @@ public class BismuthServer implements ModInitializer {
 	public static final ServerNetworking networking = new ServerNetworking();
 	public static MinecraftServer server;
 	public static DCBot bot;
-	private static final Ingredient PAPER = Ingredient.of(Items.PAPER);
-	private static final Ingredient SULPHUR = Ingredient.of(Items.GUNPOWDER);
-	private static final Recipe duration1 = new ShapelessRecipe("rocket", makeFirework(1), DefaultedList.of(Ingredient.EMPTY, PAPER, SULPHUR));
-	private static final Recipe duration2 = new ShapelessRecipe("rocket", makeFirework(2), DefaultedList.of(Ingredient.EMPTY, PAPER, SULPHUR, SULPHUR));
-	private static final Recipe duration3 = new ShapelessRecipe("rocket", makeFirework(3), DefaultedList.of(Ingredient.EMPTY, PAPER, SULPHUR, SULPHUR, SULPHUR));
-
-	static {
-		CraftingManager.register("bismuth:durationone", duration1);
-		CraftingManager.register("bismuth:durationtwo", duration2);
-		CraftingManager.register("bismuth:durationthree", duration3);
-	}
-
-	private static ItemStack makeFirework(int duration) {
-		final NbtCompound durationTag = new NbtCompound();
-		final NbtCompound fireworksTag = new NbtCompound();
-		durationTag.putByte("Flight", (byte) duration);
-		fireworksTag.put("Fireworks", durationTag);
-		final ItemStack firework = new ItemStack(Items.FIREWORKS, 3);
-		firework.setNbt(fireworksTag);
-		return firework;
-	}
 
 	@Override
     public void init() {
@@ -101,14 +71,10 @@ public class BismuthServer implements ModInitializer {
 		}
 
 		LoggerRegistry.playerConnected(player);
-		unlockCustomRecipes(player);
+		BismuthRecipeManager.unlockCustomRecipes(player);
 	}
 
 	public static void playerDisconnected(MinecraftServer server, ServerPlayerEntity player) {
 		LoggerRegistry.playerDisconnected(player);
-	}
-
-	private static void unlockCustomRecipes(ServerPlayerEntity player) {
-		player.unlockRecipes(Lists.newArrayList(duration1, duration2, duration3));
 	}
 }
