+++ src/main/resources/mixins.bismuthserver.json
@@ -7,6 +7,7 @@
 		"AnvilChunkStorageMixin",
 		"CommandManagerMixin",
 		"ConnectionListener4Mixin",
+		"CraftingManagerMixin",
 		"DirectionMixin",
 		"EndGatewayBlockEntityMixin",
 		"EntityMixin",
